pyproject.toml
[tool.poetry]
name = "aoc"
version = "0.1.0"
description = ""
authors = ["jaredculp <jared@culp.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
